[Fix] Integrate ManagerClientAccessable (caused circuler import) [Temporary Process]
